+++ Dominio/Empleado.cs
@@ -9,7 +9,7 @@ namespace Dominio
     public class Empleado
     {
         public Persona persona { get; set; }
-        public int Legajo { get; set; }
+        public long Legajo { get; set; }
         public string Contraseña { get; set; }
         public int TipoUsuario { get; set; }
         public DateTime FechaIngreso { get; set; }
